superduper: create file to log SuperDuper! message
SuperDuper! can call a script once a copy operation is complete. We use
that facility to send a message, using pushover.net to the user
informing them that the copy is complete. It isn't possible to determine
if the copy was successful, just that it has ended, which might be due
to an error/abonormal termination. A script, called pushover, is used to
send the message, and this is expected in /usr/local/bin.